Move safe_path to audeer.core.path
